+++ x/staking/migrations/v5/store.go
@@ -14,6 +14,8 @@ import (
 
 func migrateDelegationsByValidatorIndex(ctx sdk.Context, store storetypes.KVStore, cdc codec.BinaryCodec) error {
 	iterator := storetypes.KVStorePrefixIterator(store, DelegationKey)
+	iterationLimit := 1000
+	iterationCounter := 0
 
 	for ; iterator.Valid(); iterator.Next() {
 		key := iterator.Key()
@@ -23,6 +25,14 @@ func migrateDelegationsByValidatorIndex(ctx sdk.Context, store storetypes.KVStor
 		}
 
 		store.Set(GetDelegationsByValKey(val, del), []byte{})
+
+		iterationCounter++
+		if iterationCounter >= iterationLimit {
+			ctx.Logger().Info(fmt.Sprintf("Migrated 1000 delegations, next key %x", key[1:]))
+			// we set the store to the key sans the DelgationKey as we create a prefix store to iterate per block
+			store.Set(NextMigrateDelegationsByValidatorIndexKey, key[1:])
+			break
+		}
 	}
 
 	return nil